analyzer-volume_analysis/requirements.txt
numpy
pysoundfile
soundfile
wave
google-cloud-storage
flask
